feat: further extractions and cleaning
